+++ x/go/telem/frame_test.go
@@ -1080,7 +1080,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
 
@@ -1294,7 +1294,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
 
@@ -1508,7 +1508,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
 
@@ -1722,7 +1722,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
 
@@ -1936,7 +1936,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
 
@@ -2150,7 +2150,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
 
@@ -2364,7 +2364,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
 
@@ -2578,7 +2578,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
 
@@ -2792,7 +2792,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
 
@@ -3006,7 +3006,7 @@ var _ = Describe("Frame", func() {
 			series := make([]telem.Series, 256)
 			for i := range int32(256) {
 				keys[i] = i
-				series[i] = telem.NewSeriesV(int32(i), int32(i+1), int32(i+2))
+				series[i] = telem.NewSeriesV(i, i+1, i+2)
 			}
 			fr := telem.MultiFrame(keys, series)
